ENVIRONMENT: Relocated the global environment within the environment.py module.
The shortcuts get_env and reset_env are still available in
shortcuts.py, instead push_env and pop_env are only left in
environment.py . The proper way to get the environment from within
pysmt is to use pysmt.environment.get_env.
Outside of pysmt, we can simply use the shortcut.
